0410 prac done

+++ 0410확인/0410/01_pjt_template_3/articles/views.py
@@ -25,7 +25,9 @@ def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST, request.FILES)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -36,8 +38,11 @@ def create(request):
 
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
-    return redirect('articles:index')
+    if request.user.is_authenticated:
+        if request.user == article.user:
+            article.delete()
+            return redirect('articles:index')
+    return redirect('articles:detail', article.pk)
 
 
 def update(request, pk):
@@ -63,3 +68,9 @@ def comments_create(request, pk):
         comment.article = article
         comment.save()
     return redirect('articles:detail', article.pk)
+
+
+def comments_delete(request, article_pk, comment_pk):
+    comment = Comment.objects.get(pk=comment_pk)
+    comment.delete()
+    return redirect('articles:detail', article_pk)
